refactor(checker): split query helper shards
