+++ crates/tsz-checker/src/types/queries/mod.rs
@@ -15,10 +15,13 @@ pub(crate) mod binding;
 pub(crate) mod callable_truthiness;
 pub(crate) mod class;
 pub(crate) mod core;
+pub(crate) mod core_names;
 pub(crate) mod infer_bindings;
 pub(crate) mod lib;
+pub(crate) mod lib_aliases;
 pub(crate) mod lib_augmentations;
 pub(crate) mod lib_decls;
+pub(crate) mod lib_misc;
 mod lib_name_text;
 pub(crate) mod lib_namespace_direct;
 pub(crate) mod lib_prime;
